Move echoSource into targetfiltersource_test.go
